Lucene.Net.QueryParsers.Flexible.Standard.Config: Moved DateTimeOffsetUtil to Lucene.Net.Support
